requirements.txt
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cymem==2.0.6
filelock==3.4.2
flask==2.0.2
idna==3.3
itsdangerous==2.0.1
jinja2==3.0.3
langcodes==3.3.0
markupsafe==2.0.1
murmurhash==1.0.6
numpy==1.22.1
packaging==21.3
pathy==0.6.1
phonenumbers==8.12.41
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.6
pyyaml==6.0
regex==2021.11.10
requests==2.27.1
requests-file==1.5.1
six==1.16.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
tldextract==3.1.2
tqdm==4.62.3
typer==0.4.0
typing-extensions==4.0.1
urllib3==1.26.8
wasabi==0.9.0
werkzeug==2.0.2
pycryptodome==3.9.2